backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
supabase>=2.11.0,<3.0.0
pydantic>=2.10.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
